requirements.txt
# Core LLM API
google-generativeai==0.8.3

# PDF handling
PyPDF2==3.0.1
pdfminer.six==20221105

# Environment variables
python-dotenv==1.0.1

flask

faiss-cpu
pandas
numpy
scikit-learn
requests

python-docx

python-pptx

openpyxl

#embedding model
sentence_transformers

